Add type script support
