--- src/test-utils.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { ChakraProvider, theme } from '@chakra-ui/react';
-
-const AllProviders = ({ children }) => (
-  <ChakraProvider theme={theme}>{children}</ChakraProvider>
-);
-
-const customRender = (ui, options) =>
-  render(ui, { wrapper: AllProviders, ...options });
-
-export { customRender as render };
